ai-agents/video-generator-agent/requirements.txt
google-genai>=1.0.0
elevenlabs>=1.0.0
google-cloud-texttospeech>=2.14.0
moviepy>=1.0.3
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0
requests>=2.31.0
